% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_helpers.R
\name{add_color_tidyr}
\alias{add_color_tidyr}
\title{Adds color to processed tidy data}
\usage{
add_color_tidyr(x, key_values, color_fun = scales::brewer_pal(type =
  "qual", "Set1"), color_header = "#737373", color_id = "#d0d0d0")
}
\arguments{
\item{x}{a processed data-frame as outputted by process_long or process_wide}

\item{key_values}{the unique key-values}

\item{color_fun}{the color function}

\item{color_header}{the color for the header}

\item{...}{not used}
}
\value{
a data-frame with the colors
}
\description{
Adds color to processed tidy data
}
\examples{
NULL
}
