% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{anim_options}
\alias{anim_options}
\title{Animation Options}
\usage{
anim_options(transition_length = 2, state_length = 1,
  ease_default = "sine-in-out", ease_other = NULL,
  enter_ = enter_fade(), exit_ = exit_fade(),
  text_family = "Fira Sans", title_family = "Fira Mono",
  text_size = NULL, title_size = NULL, ...)
}
\arguments{
\item{transition_length}{The relative length of the transition. Will be
recycled to match the number of states in the data}

\item{state_length}{The relative length of the pause at the states. Will be
recycled to match the number of states in the data}

\item{ease_default}{Default aes easing function. See \link[tweenr:display_ease]{tweenr::display_ease}
for more options.}

\item{ease_other}{Additional aes easing options, named with aesthetic to
which the easeing should be applied, consistent with \code{\link[gganimate:ease_aes]{gganimate::ease_aes()}}.}

\item{text_family}{Font family for the plot text}

\item{title_family}{Font family for the plot title}

\item{text_size}{Font size of the plot text}

\item{title_size}{Font size of the plot title}

\item{enter_exit}{Enter/exit fading functions applied to objects in the animation.
See \link[gganimate:enter_exit]{gganimate::enter_exit} for a complete list of options.}
}
\description{
Helper function to set animation and plotting options to be passed to
\code{\link[=animate_plot]{animate_plot()}} and \code{\link[=static_plot]{static_plot()}}.
}
