% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_joins.R
\name{animate_full_join}
\alias{animate_full_join}
\title{Animates a full join}
\usage{
animate_full_join(x, y, by, export = "gif", ...)
}
\arguments{
\item{x}{the x dataset}

\item{y}{the y dataset}

\item{by}{the by arguments for the join}

\item{export}{the export type, either gif, first or last. The latter two
export ggplots of the first/last state of the join}

\item{...}{further arguments passed to base_plot}
}
\value{
either a gif or a ggplot
}
\description{
Animates a full join
}
\examples{
x <- data_frame(
  id = 1:3,
  x = paste0("x", 1:3)
)
y <- data_frame(
  id = (1:4)[-3],
  y = paste0("y", (1:4)[-3])
)

animate_full_join(x, y, by = "id", export = "first")
animate_full_join(x, y, by = "id", export = "last")

\dontrun{
  # to save the ggplot, use
  fj <- animate_full_join(x, y, by = "id", export = "last")
  ggsave("full-join.pdf", fj)

  animate_full_join(x, y, by = "id", export = "gif")
  # to save the gif, use
  fj <- animate_full_join(x, y, by = "id", export = "gif")
  anim_save(fj, "full-join.gif")
}

}
