% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_helpers.R
\name{animate_join_function}
\alias{animate_join_function}
\alias{animate_join}
\title{Animates a join - wrapper function}
\usage{
animate_join(x, y, by, type, export = "gif", ...)
}
\arguments{
\item{x}{left dataset}

\item{y}{right dataset}

\item{by}{by arguments for the join}

\item{type}{type of the join, i.e., left_join, right_join, etc.}

\item{export}{if the function exports a gif, the first, or last picture}

\item{...}{further arguments passed to static_plot or to add_color}
}
\value{
either a gif or a ggplot
}
\description{
Animates a join - wrapper function
}
\examples{
NULL
}
