% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{animate_plot}
\alias{animate_plot}
\title{Animates a plot}
\usage{
animate_plot(d, title = "", ...)
}
\arguments{
\item{d}{a preprocessed dataset}

\item{title}{the plot title}

\item{...}{further arguments passed to base_plot}
}
\value{
a gif
}
\description{
Animates a plot
}
\examples{
NULL
}
