% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_joins.R
\name{animate_semi_join}
\alias{animate_semi_join}
\title{Animates a semi join}
\usage{
animate_semi_join(x, y, by, export = "gif", ...)
}
\arguments{
\item{x}{the x dataset}

\item{y}{the y dataset}

\item{by}{the by arguments for the join}

\item{export}{the export type, either gif, first or last. The latter two
export ggplots of the first/last state of the join}

\item{...}{further arguments passed to base_plot}
}
\value{
either a gif or a ggplot
}
\description{
Animates a semi join
}
\examples{
x <- data_frame(
  id = 1:3,
  x = paste0("x", 1:3)
)
y <- data_frame(
  id = (1:4)[-3],
  y = paste0("y", (1:4)[-3])
)

animate_semi_join(x, y, by = "id", export = "first")
animate_semi_join(x, y, by = "id", export = "last")

\dontrun{
  # to save the ggplot, use
  sj <- animate_semi_join(x, y, by = "id", export = "last")
  ggsave("semi-join.pdf", sj)

  animate_semi_join(x, y, by = "id", export = "gif")
  # to save the gif, use
  sj <- animate_semi_join(x, y, by = "id", export = "gif")
  anim_save(sj, "semi-join.gif")
}

}
