% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_helpers.R
\name{animate_set_function}
\alias{animate_set_function}
\alias{animate_set}
\title{Animates a set - wrapper function}
\usage{
animate_set(x, y, type, export = "gif", ...)
}
\arguments{
\item{x}{left dataset}

\item{y}{right dataset}

\item{type}{type of the set, i.e., intersect, setdiff, etc.}

\item{export}{if the function exports a gif, the first, or last picture}

\item{...}{further arguments passed to base_plot}
}
\value{
either a gif or a ggplot
}
\description{
Animates a set - wrapper function
}
\examples{
NULL
}
