% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{base_plot}
\alias{base_plot}
\title{Prints the tiles for a processed dataset}
\usage{
base_plot(d, title = "", ...)
}
\arguments{
\item{d}{a processed dataset}

\item{title}{the title of the plot}

\item{...}{further arguments}
}
\value{
a ggplot
}
\description{
Prints the tiles for a processed dataset
}
\examples{
NULL
}
