% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combines two processed datasets and combines them for a given method}
\usage{
combine(lhs, rhs, type)
}
\arguments{
\item{lhs}{the left-hand side dataset}

\item{rhs}{the righ-hand side dataset}

\item{type}{a string of the desired combination method, allowed are all dplyr
joins or sets}
}
\value{
processed dataset of the combined values
}
\description{
Combines two processed datasets and combines them for a given method
}
\examples{
NULL
}
