% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_helpers.R
\name{dput_parser}
\alias{dput_parser}
\title{Parses a simple vector so that it looks like its input}
\usage{
dput_parser(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a string
}
\description{
Parses a simple vector so that it looks like its input
}
\examples{
dput_parser("x")
dput_parser(c("x", "y"))
}
