% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_helpers.R
\name{gather_spread}
\alias{gather_spread}
\title{Animates a gather or spread function}
\usage{
gather_spread(lhs, rhs, sequence, key_values, export, detailed, ...)
}
\arguments{
\item{lhs}{the (processed) dataset on the left-side}

\item{rhs}{the (processed) dataset on the right-side}

\item{sequence}{a named vector of the sequence titles
(current_state, final_state, operation, and reverse_operation)}

\item{key_values}{the unique key-values}

\item{export}{the export type, either gif, first or last. The latter two
export ggplots of the first/last state of the join}

\item{detailed}{boolean value if the animation should show one step for each
key value}

\item{...}{further arguments passed to animate_plot}
}
\value{
the plot or the gif
}
\description{
internally used by animate_spread and animate_gather
}
\examples{
NULL
}
