% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_helpers.R
\name{get_quos_names}
\alias{get_quos_names}
\title{Gets the ... names}
\usage{
get_quos_names(...)
}
\arguments{
\item{...}{arguments}
}
\value{
a vector of the names of ...
}
\description{
Used to get the -year
}
\examples{
x <- 1:10
y <- 1
get_quos_names(-x)
get_quos_names(x:y)
}
