% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_together.R
\name{move_together}
\alias{move_together}
\title{Combines two processed datasets and combines them for a given method}
\usage{
move_together(lhs, rhs, type)
}
\arguments{
\item{lhs}{the left-hand side dataset}

\item{rhs}{the righ-hand side dataset}

\item{type}{a string of the desired combination method, allowed are all dplyr
joins or sets}
}
\value{
processed dataset of the combined values
}
\description{
Combines two processed datasets and combines them for a given method
}
\examples{
NULL
}
