% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_helpers.R
\name{process_join}
\alias{process_join}
\title{Preprocess data}
\usage{
process_join(x, y, by, fill = TRUE, ...)
}
\arguments{
\item{x}{a left dataset}

\item{y}{a right dataset}

\item{by}{a by argument for joins / set operations}

\item{fill}{if missing ids should be filled}

\item{...}{further arguments passed to add_color}
}
\value{
a preprocessed dataset
}
\description{
Preprocess data
}
\examples{
NULL
test for
a <- c("unique", "mult", "mult", "also unique")
add_duplicate_number(a)
}
