% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_helpers.R
\name{process_long}
\alias{process_long}
\title{Processes a long dataframe and converts it into a dataset that can be plotted}
\usage{
process_long(x, ids, key, value, color_id = "lightgray", ...)
}
\arguments{
\item{x}{a long data frame}

\item{ids}{a vector of id-variables that are already in the tidy-format}

\item{key}{a vector of key-variables}

\item{color_id}{the color for the id-body}

\item{...}{}
}
\value{

}
\description{
Processes a long dataframe and converts it into a dataset that can be plotted
}
\examples{
long <- data_frame(
  year = c(2010L, 2011L, 2010L, 2011L, 2010L, 2011L),
  person = c("Alice", "Alice", "Bob", "Bob", "Charlie", "Charlie"),
  sales = c(105, 110, 100, 97, 90, 95)
)
process_long(long, ids = "year", key = "person", value = "sales")
process_long(long, ids = "year", key = "person", value = "sales") \%>\% static_plot
}
