% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyr_helpers.R
\name{process_wide}
\alias{process_wide}
\title{Processes a wide dataframe and converts it into a dataset that can be plotted}
\usage{
process_wide(x, ids, key, color_id = "lightgray", ...)
}
\arguments{
\item{x}{a wide data frame}

\item{ids}{a vector of id-variables that are already in the tidy-format}

\item{key}{a vector of key-variables}

\item{color_id}{the color for the id-body}

\item{...}{}
}
\value{

}
\description{
Processes a wide dataframe and converts it into a dataset that can be plotted
}
\examples{
wide <- data_frame(
  year = 2010:2011,
  Alice = c(105, 110),
  Bob = c(100, 97),
  Charlie = c(90, 95)
)
process_wide(wide, ids = "year", key = "person")
process_wide(wide, ids = "year", key = "person") \%>\% static_plot
}
