% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{static_plot}
\alias{static_plot}
\title{Prints the tiles for a processed dataset statically}
\usage{
static_plot(d, title = "", ..., anim_opts = anim_options(...))
}
\arguments{
\item{d}{a processed dataset}

\item{title}{the title of the plot}

\item{...}{Arguments passed on to \code{anim_options}
\describe{
  \item{text_family}{Font family for the plot text}
  \item{title_family}{Font family for the plot title}
  \item{text_size}{Font size of the plot text}
  \item{title_size}{Font size of the plot title}
  \item{ease_default}{Default aes easing function. See \code{\link[tweenr:display_ease]{tweenr::display_ease()}}
for more options.}
  \item{ease_other}{Additional aes easing options, specified as a named list.
List entries are named with the aesthetic to which the easeing should be
applied, consistent with \code{\link[gganimate:ease_aes]{gganimate::ease_aes()}}.
E.g. \code{list(color = "sine")}.}
  \item{enter}{Enter fading function applied to objects in the animation. See
\link[gganimate:enter_exit]{gganimate::enter_exit} for a complete list of options.}
  \item{exit}{Exit fading function applied to objects in the animation. See
\link[gganimate:enter_exit]{gganimate::enter_exit} for a complete list of options.}
  \item{transition_length}{The relative length of the transition. Will be
recycled to match the number of states in the data}
  \item{state_length}{The relative length of the pause at the states. Will be
recycled to match the number of states in the data}
}}

\item{anim_opts}{Animation options generated with \code{\link[=anim_options]{anim_options()}}. Overrides
any options set in \code{...}.}
}
\value{
a ggplot
}
\description{
Prints the tiles for a processed dataset statically
}
\examples{
NULL
}
