% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz-package.R
\docType{package}
\name{tidyexplain-package}
\alias{tidyexplain}
\alias{tidyexplain-package}
\title{tidyexplain: Animated Explanations of Tidyverse Verbs}
\description{
Animated explanations of the verbs in the tidyverse
    using gganimate and ggplot2.
}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{g.adenbuie@gmail.com}

Authors:
\itemize{
  \item David Zimmermann \email{david_j_zimmermann@hotmail.com}
}

Other contributors:
\itemize{
  \item Tyler Grant Smith [contributor]
}

}
\keyword{internal}
