% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz-package.R
\docType{package}
\name{tidyverbs-package}
\alias{tidyverbs}
\alias{tidyverbs-package}
\title{tidyverbs: Animate the Verbs of the Tidyverse}
\description{
Animate the verbs of the tidyverse.
}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{g.adenbuie@gmail.com}

Authors:
\itemize{
  \item David Zimmermann \email{david_j_zimmermann@hotmail.com}
}

Other contributors:
\itemize{
  \item Tyler Grant Smith [contributor]
}

}
\keyword{internal}
