% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_helpers.R
\name{add_color}
\alias{add_color}
\title{Adds Color to a processed data_frame}
\usage{
add_color(x, ids, by, color_header = "#737373",
  color_other = "#d0d0d0", color_missing = "#ffffff",
  color_fun = scales::brewer_pal(type = "qual", "Set1"),
  text_color = NA, ...)
}
\arguments{
\item{x}{a processed data_frame}

\item{ids}{a vector of ids for the color-matching}

\item{by}{a vector of column names that constitute the by-argument of joins/sets}

\item{color_header}{color for the header}

\item{color_other}{color for "inactive" values}

\item{color_missing}{color for missing values}

\item{color_fun}{the function to generate the colors}

\item{text_color}{the color for the text inside the tiles,
defaults to white/black depending on tile color}

\item{...}{}
}
\value{
the processed data_frame with a new column .color
}
\description{
Adds Color to a processed data_frame
}
\examples{
NULL
}
