% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{animate_plot}
\alias{animate_plot}
\title{Animates a plot}
\usage{
animate_plot(d, title = "", transition_length = 2, state_length = 1,
  ...)
}
\arguments{
\item{d}{a preprocessed dataset}

\item{title}{the plot title}

\item{transition_length}{see transition_states}

\item{state_length}{see transition_states}

\item{...}{further arguments passed to static_plot}
}
\value{
a gif
}
\description{
Animates a plot
}
\examples{
NULL
}
