% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_helpers.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data}
\usage{
preprocess_data(x, y, by)
}
\arguments{
\item{x}{a left dataset}

\item{y}{a right dataset}

\item{by}{a by argument for joins / set operations}
}
\value{
a preprocessed dataset
}
\description{
Preprocess data
}
\examples{
NULL
}
