% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_helpers.R
\name{process_data}
\alias{process_data}
\title{Processes the data}
\usage{
process_data(x, ids, by, width = 1, side = NA, fill = TRUE)
}
\arguments{
\item{x}{a preprocessed dataset}

\item{ids}{a data_frame of ids (.id and .id_long)}

\item{by}{a vector of by-arguments}

\item{width}{the width of the tiles}

\item{side}{the side (x or y, lhs or rhs, etc)}

\item{fill}{if missing ids should be filled}
}
\value{
a data_frame including all necessary information
}
\description{
Processes the data
}
\examples{
NULL
}
