% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{static_plot}
\alias{static_plot}
\title{Prints the tiles for a processed dataset statically}
\usage{
static_plot(d, title = "", text_family = "Fira Sans",
  title_family = "Fira Mono", text_size = 5, title_size = 17, ...)
}
\arguments{
\item{d}{a processed dataset}

\item{title}{the title of the plot}

\item{text_family}{the font for the text}

\item{title_family}{the font for the title}

\item{text_size}{the size of the text}

\item{title_size}{the size of the title}

\item{...}{further arguments}
}
\value{
a ggplot
}
\description{
Prints the tiles for a processed dataset statically
}
\examples{
NULL
}
